Moved ExecuteOptions to a new package
